+++ balances/get_detailed_balance.js
@@ -2,11 +2,15 @@ const asyncAuto = require('async/auto');
 const {formatTokens} = require('ln-sync');
 const {getChainTransactions} = require('ln-service');
 const {getChannels} = require('ln-service');
+const {getLockedUtxos} = require('ln-service');
 const {getPendingChannels} = require('ln-service');
 const {getUtxos} = require('ln-service');
 const {returnResult} = require('asyncjs-util');
+const {Transaction} = require('bitcoinjs-lib');
 
 const detailedBalances = require('./detailed_balances');
+
+const {fromHex} = Transaction;
 
 /** Get a detailed balance that categorizes balance of tokens on the node
 
@@ -39,6 +43,18 @@ module.exports = (args, cbk) => {
         return getChannels({lnd: args.lnd}, cbk);
       }],
 
+      // Get locked UTXOs
+      getLocked: ['validate', ({}, cbk) => {
+        return getLockedUtxos({lnd: args.lnd}, (err, res) => {
+          // Ignore errors
+          if (!!err) {
+            return cbk(null, []);
+          }
+
+          return cbk(null, res.utxos);
+        });
+      }],
+
       // Get pending channels
       getPending: ['validate', ({}, cbk) => {
         return getPendingChannels({lnd: args.lnd}, cbk);
@@ -52,17 +68,49 @@ module.exports = (args, cbk) => {
       // Get the UTXOs
       getUtxos: ['validate', ({}, cbk) => getUtxos({lnd: args.lnd}, cbk)],
 
+      // Cross reference locked transactions to UTXO data
+      locked: ['getLocked', 'getTx', ({getLocked, getTx}, cbk) => {
+        const {transactions} = getTx;
+
+        const utxos = getLocked.map(locked => {
+          // Exit early when the lock is expired
+          if (locked.lock_expires_at < new Date().toISOString()) {
+            return;
+          }
+
+          const tx = transactions.find(n => n.id === locked.transaction_id);
+
+          // Exit early when there is no related confirmed transaction
+          if (!tx || !tx.transaction || !tx.confirmation_count) {
+            return;
+          }
+
+          const output = fromHex(tx.transaction).outs[locked.transaction_vout];
+
+          // Exit early when the output is not found in the transaction
+          if (!output) {
+            return;
+          }
+
+          return {tokens: output.value};
+        });
+
+        return cbk(null, utxos.filter(n => !!n));
+      }],
+
       // Calculate balance
       balance: [
         'getChannels',
         'getPending',
         'getTx',
         'getUtxos',
-        ({getChannels, getPending, getTx, getUtxos}, cbk) =>
+        'locked',
+        ({getChannels, getPending, getTx, getUtxos, locked}, cbk) =>
       {
         const format = tokens => formatTokens({tokens}).display.trim();
 
         const balances = detailedBalances({
+          locked,
           channels: getChannels.channels,
           pending: getPending.pending_channels,
           transactions: getTx.transactions,